pyproject.toml
[project]
name = "langchain-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.4",
]

[tool.pyright]
venvPath="."
venv = ".venv"

[tool.basedpyright]
venvPath="."
venv = ".venv"
